flask_yolov8_api/requirements.txt.txt
flask
ultralytics
Pillow